Implemented GPS data parsing
