+++ Core/Inc/processing.h
@@ -1,0 +1,7 @@
+#ifndef __PROCESSING_H__
+#define __PROCESSING_H__
+#include "main.h"
+#include"location.h"
+#endif
+
+void processMessage(char *str);
