+++ Core/Inc/sms.h
@@ -1,0 +1,11 @@
+#ifndef __SMS_H__
+#define __SMS_H__
+#include "main.h"
+#endif
+
+void checkSMS(void);
+
+
+int SMS;
+
+extern int SMS;
